backend/requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
whitenoise
django-cors-headers
python-dotenv